# SearchTagValuesV2 use protobuf internally instead of json to reduce latency