**What this PR does**:
The frontend communicates with the queriers using json, and this is becomes a bottleneck for large autocomplete lookups.  This PR switches /api/v2/search/tag/.../values to use http/protobuf instead.  Testing locally and in a larger cell on a large lookup, the savings are around 30% user-facing latency to Grafana. Timings:  15s -> 10s, 5s -> 3s.   We like graphs, so here is one showing the second set:

<img width="1882" alt="image" src="https://github.com/grafana/tempo/assets/13524475/e2092847-9665-42d9-8334-4e516f7d4b0e">


The encoding is implemented optionally where the querier will reply with proto only if the request header is `Accept: application/protobuf`.  Else it defaults to json.  This helps with compatibility during rollouts of old/new pods, and also curl-ability. 

A couple future steps:

(1) It might be worth swapping over the remaining apis with large payloads like search/tags and metrics.  But didn't want to expand scope.

(2) Even better would be preserve the protobuf all the way to Grafana like we do for trace lookup, but there is work needed to connect the response/request through the new async pipeline in the frontend, so the final response can be built with the appropriate encoding.

**Which issue(s) this PR fixes**:
Fixes n/a

**Checklist**
- [ ] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`